refactor: code for add row sheet
